requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
pandas
google-generativeai
python-dotenv
requests
streamlit
plotly